pyproject.toml
[tool.poetry]
name = "hackadoo"
version = "0.1.0"
description = "MIT"
authors = ["Minigrim0, Morti, ZoroxaT, Nico"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
requests = "^2.28.2"
wikipedia-api = "^0.5.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
